+++ relay-server/src/envelope/item.rs
@@ -456,14 +456,24 @@ impl Item {
     }
 
     /// Sets the parent entity that this item is associated with.
-    pub fn set_parent_id(&mut self, parent_id: ParentId) {
-        self.headers.parent_id = Some(parent_id);
+    pub fn set_parent_id(&mut self, parent_id: Option<ParentId>) {
+        self.headers.parent_id = parent_id;
     }
 
     /// Returns `true` if this item is an attachment with AttachmentV2 content type.
-    pub fn is_attachment_v2(&self) -> bool {
+    fn is_attachment_v2(&self) -> bool {
         self.ty() == &ItemType::Attachment
-            && self.content_type() == Some(&ContentType::AttachmentV2)
+            && self.content_type() == Some(&ContentType::TraceAttachment)
+    }
+
+    /// Returns `true` if this item is a V2 attachment owned by spans.
+    pub fn is_span_attachment(&self) -> bool {
+        self.is_attachment_v2() && matches!(self.parent_id(), Some(ParentId::SpanId(_)))
+    }
+
+    /// Returns `true` if this item is a V2 attachment without any span/log/etc. association.
+    pub fn is_trace_attachment(&self) -> bool {
+        self.is_attachment_v2() && self.parent_id().is_none()
     }
 
     /// Returns the attachment payload size.
@@ -998,13 +1008,13 @@ pub struct ItemHeaders {
 
     /// Content length of an optional meta segment that might be contained in the item.
     ///
-    /// For the time being such an meta segment is only present for span attachments.
+    /// For the time being such an meta segment is only present for trace attachments.
     #[serde(skip_serializing_if = "Option::is_none")]
     meta_length: Option<u32>,
 
     /// Parent entity that this item is associated with, if any.
     ///
-    /// For the time being only applicable if the item is a span-attachment.
+    /// For the time being only applicable if the item is a trace attachment.
     #[serde(flatten, skip_serializing_if = "Option::is_none")]
     parent_id: Option<ParentId>,