+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 - Increases log size limits to better support log drains. ([#5441](https://github.com/getsentry/relay/pull/5441))
 - Add functionality to process and store span attachments. ([#5423](https://github.com/getsentry/relay/pull/5423), [#5363](https://github.com/getsentry/relay/pull/5363))
 - Apply existing cookie rules to `http.request.header.cookie.<key>` fields. ([#5456](https://github.com/getsentry/relay/pull/5456))
+- Add functionality to process and store trace attachments. ([#5457](https://github.com/getsentry/relay/pull/5457))
 - Lower default memory utilization threshold for disk spooling from 90% to 80%. When Relay hits 80% memory utilization, it will start spooling envelopes to disk instead of processing them. ([#5472](https://github.com/getsentry/relay/pull/5472))
+
 
 **Bug Fixes**: